Massive API overhaul, plus some cleanup.
This really should have gone into separate commits. I didn't expect it to explode so quickly =/

The existing API has evolved into something far more complicated than it should be. My first stab at containing the chaos is to add another layer on top of everything to give the end user a simpler interface to work from (users should never touch filters and sinks in a normal world). Hopefully I can reorganize some of the guts to slim things down afterwards. We'll see…

The new end-user interface is the "Installation". This gives a kind of clean-slate class template to use for whatever backend system you're targeting. I've made installations for Quincy and Hockey, and maybe I'll even get off my ass and write for other backends as well! Chances that writing other backend adaptors will complicate the API: About 100%, but that's a battle for another day.

I've also changed KSCrash to use properties for configuration rather than init values. You can set any property at any time without worry. I've also changed the default operation mode to be more conservative (no zombie tracking or deadlock detection by default).

As well, KSCrash will not attempt to send crash reports to a backend server unless you explicitly tell it to. Sending automatically is too restrictive for many use cases.

The rest of the changes are code cleanup (using __unused instead of #pragma unused, removing unnecessary includes, naming, DRY, stuff like that).

+++ KSCrash/KSCrash/KSCrashSentry_Deadlock.h
@@ -56,7 +56,7 @@ void kscrashsentry_uninstallDeadlockHandler(void);
  *
  * @param value The number of seconds between checks (0 = disabled).
  */
-void kscrashSentry_setDeadlockHandlerWatchdogInterval(float value);
+void kscrashSentry_setDeadlockHandlerWatchdogInterval(double value);
 
 
 #ifdef __cplusplus
